Removed tests from the MathFunctionsTests class and moved them to their respective test classes. Also brought Log class up to coding standard.
